Add a resolution hint that tells you what to write to .bazelifyrc
This makes it a lot easier to fix resolution issues. It takes any unresolved includes, and adds a number of target overrides to the existing .bazelifyrc where the unresolved targets show as "PLEASE RESOLVE: possible0|possible1|possible2". If a user writes this to their .bazelifyrc and fixes the PLEASE REMOVE lines, then the config should work.