Improve resolution hint with override safeguards and write to file
The original output was very verbose. Instead, we now write to the .bazelifyrc.hint file. This is nicer because it lets the user simply move the file ove, and the stdout of nrfbazelify is a lot smaller (and readable!). In addition, it lets us test the hint output.

Also add a test for resolution hint output.