Improve hint by including what requires each unresolved target
I added a REQUIRED BY x,y,z in front of the PLEASE RESOLVE message. x,y,z are file names that include this target. In verbose mode, it will print the pretty SDK path, which is relative to the sdk directory. Otherwise, it will print just the file name that includes the target.