Return an error if loading .bazelifyrc fails (#20)
I originally skipped .bazelifyrc failures, because it's possible to not have a .bazelifyrc. But, I was also opaquely skipping proto marshalling errors, which led to some confusion. Now, I Stat the file, and print a loud WARNING if there's no file. For other errors, I throw an actual error and exit early, instead of continuing.