Create a new Label library which handles formatting
This makes it possible to separate label creation from formatting. We need this later to handle moving labels. As-is, if a label is created as a relative label, and it gets moved, we lose information about its directory.

This also puts all the label formatting code in one place, which is much cleaner.